% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_sector.R
\name{bridge_sector}
\alias{bridge_sector}
\title{Add the column \code{sector} (as classified by 2dii) and \code{borderline}}
\usage{
bridge_sector(data)
}
\arguments{
\item{data}{A loanbook dataframe.}
}
\value{
A loanbook dataframe with additional \code{sector} and \code{borderline}
columns.
}
\description{
This function adds two columns, \code{sector} and \code{borderline}:
\itemize{
\item \code{sector}: \code{sector} gives climate-relevant sectors looked up from standard
sector classifications in production databases.
\item \code{borderline}: Indicates if the classification is borderline or not.
}
}
\details{
A company acts in a climate-relevant sector if it meets two conditions:
\enumerate{
\item It is classified in one of the sectors that translate 1 to 1 from a a
standard sector classification to a climate-relevant sector covered by the
PACTA analysis (e.g. utilities to power).
\item It is classified in a sector that can but does not necessarily map to a
climate-relevant sector covered by the PACTA analysis (e.g. power
grid/network to Power, general mining to coal mining).
}

For a list of sector classification systems see
\link[r2dii.dataraw:classification_bridge]{r2dii.dataraw::classification_bridge}.
}
\examples{
# Must be attached
library(r2dii.dataraw)

bridge_sector(r2dii.dataraw::loanbook_demo)
}
\seealso{
Other internal-ish: 
\code{\link{replace_customer_name}()},
\code{\link{string_similarity}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
