% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_match_success_rate.R
\name{calculate_match_success_rate}
\alias{calculate_match_success_rate}
\title{Calculate the success rate of the matching process}
\usage{
calculate_match_success_rate(matched, loanbook)
}
\arguments{
\item{matched}{A data frame like the validated output of \code{\link[=match_name]{match_name()}}.}

\item{loanbook}{A data frame structured like \link[r2dii.data:loanbook_demo]{r2dii.data::loanbook_demo}.}
}
\value{
A data frame.
}
\description{
Calculate the success rate of the matching process
}
\examples{
matched <- match_name(r2dii.data::loanbook_demo, r2dii.data::abcd_demo)
calculate_match_success_rate(matched, r2dii.data::loanbook_demo)

}
\seealso{
Other main functions: 
\code{\link{match_name}()},
\code{\link{prioritize}()}
}
\concept{main functions}
