% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crucial_lbk.R
\name{crucial_lbk}
\alias{crucial_lbk}
\title{Crucial \code{loanbook} columns for \code{match_name()}}
\usage{
crucial_lbk()
}
\value{
A character vector.
}
\description{
This is a helper to select the minimum \code{loanbook} columns you need to run
\code{match_name()}. Using more columns may use too much time and memory.
}
\examples{
crucial_lbk()
}
\seealso{
Other helpers: 
\code{\link{prioritize_level}()}
}
\concept{helpers}
