% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_addin.R
\name{match_addin}
\alias{match_addin}
\alias{match_app}
\title{Match loanbook with asset level data}
\usage{
match_addin()

match_app()
}
\value{
Called for its side effects.
}
\description{
FIXME: This is work in progress. At this stage these functions do useless
stuff. We use them only to test the interfaces they provide:
\itemize{
\item \code{match_addin()}:
\itemize{
\item The user enters parameters on a pop-up window.
\item The app prints stuff on the R console.
\item It must run from R or RStudio (can't be hosted online).
}
\item \code{match_app()}:
\itemize{
\item The user enters parameters on the web browser.
\item The app prints stuff on the web browser.
\item It can be hosted online, maybe on our private server.
}
}
}
\examples{
\dontrun{
match_addin()

match_app()
}
}
