% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matching.R
\name{match_all_against_all}
\alias{match_all_against_all}
\title{Run matching between two lists}
\usage{
match_all_against_all(
  names_to_match,
  names_to_match_against,
  threshold = 0.95,
  ...
)
}
\arguments{
\item{names_to_match}{List of names to match.}

\item{names_to_match_against}{List of names to be matched against.}

\item{threshold}{A threshold matching score to reach}

\item{...}{Additional arguments are passed on to \link[stringdist:stringsim]{stringdist::stringsim}.}
}
\value{
Returns a tibble of all matching name combinations scoring above
the threshold.
}
\description{
Applies the string_similarity function to every combination of two lists:
\code{names_to_match} and \code{names_to_match_against}, meeting a minimum threshold
outputting the test_name, matched_name and score.
}
\examples{
library(r2dii.dataraw)

loanbook_entities <- loanbook_demo \%>\%
  prepare_loanbook_for_matching()

ald_entities <- ald_demo \%>\%
  prepare_ald_for_matching() \%>\%
  dplyr::mutate(simpler_name = replace_customer_name(.data$name))

match_all_against_all(loanbook_entities$simpler_name,
             ald_entities$simpler_name)
}
