% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{overwrite_name_sector}
\alias{overwrite_name_sector}
\title{Overwrite the name and/ or sector of prepared loanbook data}
\usage{
overwrite_name_sector(data, overwrite)
}
\arguments{
\item{data}{A matching-prepared loanbook}

\item{overwrite}{A dataframe used to overwrite the sector and/or name of a
particular direct loantaker or ultimate parent. If only name (sector)
should be overwritten leave sector (name) as \code{NA}.}
}
\value{
A dataframe of all unique name + sector combinations, including all
IDs, and with elements already manually overwritten.
}
\description{
This function overwrites the name and/ or sector of entities of the perpared
loanbook as specified by manual input. Entities to be overwritten are
identified by level (e.g. \code{direct_loantaker} or \code{ultimate_parent}) and ID.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  bridge_sector() \%>\%
  prepare_loanbook_for_matching() \%>\%
  overwrite_name_sector(overwrite_demo)
}
\seealso{
\link[r2dii.dataraw:overwrite_demo]{r2dii.dataraw::overwrite_demo} demo overwrite input file.
}
