% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{prepare_ald_for_matching}
\alias{prepare_ald_for_matching}
\alias{prepare_loanbook_for_matching}
\title{Prepare loanbook and asset level data (ald) for matching}
\usage{
prepare_ald_for_matching(data)

prepare_loanbook_for_matching(data, overwrite = NULL)
}
\arguments{
\item{data}{A dataframe. Should be a loanbook or asset level dataset.}

\item{overwrite}{A dataframe used to overwrite the sector and/or name of a
particular direct loantaker or ultimate parent. If only name (sector)
should be overwritten leave sector (name) as \code{NA}.}
}
\value{
A dataframe of all unique name + sector combinations, including all
IDs, and with elements already manually overwritten.
}
\description{
These functions take all unique \code{name} + \code{sector} combinations of a loanbook
or asset level data, preparing it for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  prepare_loanbook_for_matching()

loanbook_demo \%>\%
  prepare_loanbook_for_matching(overwrite = overwrite_demo)

ald_demo \%>\%
  prepare_ald_for_matching()
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description}, \link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.

Other functions to prepare data for matching.: 
\code{\link{bridge_sector}()}
}
\concept{functions to prepare data for matching.}
