% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize}
\alias{prioritize}
\title{Pick rows where \code{score} is 1 and \code{level} per loan is of highest \code{priority}}
\usage{
prioritize(data, priority = NULL)
}
\arguments{
\item{data}{A  dataframe, commonly the output of \code{\link[=match_name]{match_name()}}.}

\item{priority}{One of:
\itemize{
\item \code{NULL}: defaults to the default level priority as returned by
\code{\link[=prioritize_level]{prioritize_level()}}.
\item A character vector giving a custom priority.
\item A function to apply to the output of \code{\link[=prioritize_level]{prioritize_level()}}, e.g. \code{rev}.
\item A quosure-style lambda function, e.g. \code{~ rev(.x)}.
}}
}
\value{
A dataframe with a single row per loan, where \code{score} is 1 and
priority level is highest.
}
\description{
Pick rows where \code{score} is 1 and \code{level} per loan is of highest \code{priority}
}
\section{Handling grouped data}{

This function ignores but preserves existing groups them.
}

\examples{
library(dplyr)

matched <- tribble(
  ~sector, ~sector_ald, ~score,  ~id,  ~level,
  "coal",  "coal",      1,       "aa", "ultimate_parent",
  "coal",  "coal",      1,       "aa", "direct_loantaker",
  "coal",  "coal",      1,       "bb", "intermediate_parent",
  "coal",  "coal",      1,       "bb", "ultimate_parent",
)

prioritize_level(matched)

# Using default priority
prioritize(matched)

# Using the reverse of the default priority
prioritize(matched, priority = rev)

# Same
prioritize(matched, priority = ~ rev(.x))

# Using a custom priority
bad_idea <- c("intermediate_parent", "ultimate_parent", "direct_loantaker")

prioritize(matched, priority = bad_idea)
}
\seealso{
\code{\link[=match_name]{match_name()}}, \code{\link[=prioritize_level]{prioritize_level()}}.

Other user-oriented: 
\code{\link{match_name}()}
}
\concept{user-oriented}
