% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_priority.R
\name{prioritize_at}
\alias{prioritize_at}
\title{Pick rows from a dataframe based on a priority set at some columns}
\usage{
prioritize_at(data, .at, priority = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{.at}{Most commonly, a character vector of one column name. For more
general usage see the \code{.vars} argument to \code{\link[dplyr:arrange_at]{dplyr::arrange_at()}}.}

\item{priority}{Most commonly, a character vector of the priority to
re-order the column(x) given by \code{.at}.}
}
\value{
A dataframe, commonly with less rows than the input.
}
\description{
Pick rows from a dataframe based on a priority set at some columns
}
\examples{
library(dplyr)

data <- tibble::tribble(
  ~x, ~y,
  1, "a",
  2, "a",
  2, "z",
)

data \%>\% prioritize_at("y")

data \%>\%
  group_by(x) \%>\%
  prioritize_at("y")

data \%>\%
  group_by(x) \%>\%
  prioritize_at(.at = "y", priority = c("z", "a")) \%>\%
  arrange(x) \%>\%
  ungroup()
}
