% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize_level}
\alias{prioritize_level}
\title{Unique level values ordered according to the default level priority}
\usage{
prioritize_level(data)
}
\arguments{
\item{data}{A dataframe, commonly the output of \code{\link[=match_name]{match_name()}}.}
}
\value{
A character vector of the default level priority per loan.
}
\description{
Unique level values ordered according to the default level priority
}
\examples{
matched <- tibble::tibble(
  level_lbk = c(
    "intermediate_parent_1",
    "direct_loantaker",
    "direct_loantaker",
    "direct_loantaker",
    "ultimate_parent",
    "intermediate_parent_2"
  )
)
prioritize_level(matched)
}
\seealso{
Other internal-ish: 
\code{\link{bridge_sector}()},
\code{\link{match_all_against_all}()},
\code{\link{prioritize_at}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_similarity}()},
\code{\link{select_chr}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
