% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unique_name_sector_combinations.R
\name{pull_unique_name_sector_combinations}
\alias{pull_unique_name_sector_combinations}
\title{Prepare the loanbook elements to be matched.}
\usage{
pull_unique_name_sector_combinations(data)
}
\arguments{
\item{data}{A dataframe structured like the output of \code{\link[=bridge_sector]{bridge_sector()}}.}
}
\value{
A dataframe of all unique name + sector combinations, including all
IDs, and with elements already manually overwritten.
}
\description{
This function takes all unique name + sector combinations of a loanbook,
preparing it for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  bridge_sector() \%>\%
  pull_unique_name_sector_combinations()
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description}.

Other functions to prepare data for matching.: 
\code{\link{bridge_sector}()},
\code{\link{overwrite_name_sector}()},
\code{\link{prepare_loanbook_for_matching}()}
}
\concept{functions to prepare data for matching.}
