% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unique_name_sector_combinations_ald.R
\name{pull_unique_name_sector_combinations_ald}
\alias{pull_unique_name_sector_combinations_ald}
\title{Prepare the ald elements to be matched.}
\usage{
pull_unique_name_sector_combinations_ald(data)
}
\arguments{
\item{data}{A dataframe structured like \link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.}
}
\value{
A dataframe of all unique name + sector combinations.
}
\description{
This function takes all unique name + sector combinations of the ald,
preparing it for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

ald_demo \%>\%
  pull_unique_name_sector_combinations_ald()
}
\seealso{
\link[r2dii.dataraw:ald_description]{r2dii.dataraw::ald_description}.

Other functions to prepare data for matching.: 
\code{\link{bridge_sector}()},
\code{\link{overwrite_name_sector}()},
\code{\link{prepare_ald_for_matching}()},
\code{\link{prepare_loanbook_for_matching}()},
\code{\link{pull_unique_name_sector_combinations}()}
}
\concept{functions to prepare data for matching.}
