% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.match-package.R
\docType{package}
\name{r2dii.match-package}
\alias{r2dii.match}
\alias{r2dii.match-package}
\title{r2dii.match: Tools to Match a Loanbook with Asset Level Data}
\description{
Tools to match a loanbook with asset level data.
    For more information visit <https://2degrees-investing.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.match}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.match/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Authors:
\itemize{
  \item Klaus Hagedorn \email{klaus@2degrees-investing.org}
  \item George Harris \email{george@2degrees-investing.org}
  \item Jackson Hoffart \email{jackson@2degrees-investing.org} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID})
  \item Daan Koopman \email{daan@2degrees-investing.org}
  \item Florence Palandri \email{florence@2degrees-investing.org}
  \item Evgeny Petrovsky [contractor]
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
