% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_customer_name.R
\name{replace_customer_name}
\alias{replace_customer_name}
\title{Simplify customer name}
\usage{
replace_customer_name(x, from_to = NULL, ownership = NULL,
  remove_ownership = FALSE)
}
\arguments{
\item{x}{Character string, (likely) containing customer names.}

\item{from_to}{A dataframe with replacement rules to be applied, contains
columns \code{from} (for initial values) and \code{to} (for resulting values).}

\item{ownership}{vector of company ownership types to be distinguished for
cut-off or separation.}

\item{remove_ownership}{Flag that defines whether ownership type (like llc)
should be cut-off.}
}
\value{
Character string.
}
\description{
Although this function takes any character vector, it is useful for a vector
containing customer names. This is what this function does to it:
\itemize{
\item Convert to lower case;
\item Convert to latin-ascii characters;
\item Convert ownership types to standard abbreviations.
}
}
\examples{
replace_customer_name("A. and B")
replace_customer_name("Acuity Brands Inc")
replace_customer_name(c("3M Company", "Abbott Laboratories", "AbbVie Inc."))

custom_replacement <- tibble::tibble(from = "AAAA", to = "B")
replace_customer_name("Aa Aaaa", from_to = custom_replacement)

neutral_replacement <- tibble::tibble(from = character(0), to = character(0))
replace_customer_name("Company Name Owner", from_to = neutral_replacement)
replace_customer_name(
  "Company Name Owner",
  from_to = neutral_replacement,
  ownership = "owner",
  remove_ownership = TRUE
)
}
\author{
person(given = "Evgeny",
family = "Petrovsky",
role = c("aut", "ctr"))
}
