% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{restructure_ald_for_matching}
\alias{restructure_ald_for_matching}
\title{Prepares asset-level data (ald) for the fuzzy matching process}
\usage{
restructure_ald_for_matching(data)
}
\arguments{
\item{data}{A dataframe. Should be an asset-level dataset.}
}
\value{
A dataframe with unique combinations of \code{name} + \code{sector}, including
all IDs, and with elements already manually overwritten.
}
\description{
This function prepares asset-level data (ald) for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

ald_demo \%>\%
  restructure_ald_for_matching()
}
\seealso{
\link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.

Other user-oriented: 
\code{\link{match_all_against_all}()},
\code{\link{restructure_loanbook_for_matching}()}
}
\concept{user-oriented}
