% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_loanbook_for_matching.R
\name{restructure_loanbook_for_matching}
\alias{restructure_loanbook_for_matching}
\title{Restructure  loanbook dataset (lbk) in preparation for fuzzy matching}
\usage{
restructure_loanbook_for_matching(data, overwrite = NULL)
}
\arguments{
\item{data}{A dataframe. Should be a loanbook.}

\item{overwrite}{A dataframe used to overwrite the \code{sector} and/or \code{name}
columns of a particular direct loantaker or ultimate parent. To overwrite
only \code{sector}, the value in the \code{name} column should be \code{NA}.}
}
\value{
A dataframe with unique combinations of \code{name} + \code{sector}, including
all IDs, and with elements already manually overwritten.
}
\description{
This function restructures a loanbook dataset (lbk) in preparation for
the fuzzy matching process. Most notably, it outputs an the \code{alias} column
from values in the columns \code{name_direct_loantaker} and
\code{name_ultimate_parent}.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  restructure_loanbook_for_matching()

loanbook_demo \%>\%
  restructure_loanbook_for_matching(overwrite = overwrite_demo)
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description}

Other internal-ish: 
\code{\link{add_sector_and_borderline}()},
\code{\link{prioritize_at}()},
\code{\link{prioritize_level}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{score_alias_similarity}()},
\code{\link{score_string_similarity}()},
\code{\link{select_chr}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
