% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_simpler_name_by_sector.R
\name{score_simpler_name_by_sector}
\alias{score_simpler_name_by_sector}
\title{Score similarity between \code{simpler_name} values in two dataframes, by sector}
\usage{
score_simpler_name_by_sector(x, y)
}
\arguments{
\item{x, y}{Dataframes}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Apply \code{string_similarity()} to all combinations of \code{simpler_name} values, by
sector, from two dataframes.
}
\examples{
library(dplyr)

x <- tibble(sector = c("A", "B", "B"), simpler_name = c("xa", "xb", "xc"))
y <- tibble(sector = c("A", "B", "C"), simpler_name = c("ya", "yb", "yc"))

scores <- score_simpler_name_by_sector(x, y)
scores

scores \%>\%
  dplyr::filter(score > 0.5)
}
