% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{simplify_name_column}
\alias{simplify_name_column}
\title{Wrapper to simplify the name column of the prepared loanbook}
\usage{
simplify_name_column(data)
}
\arguments{
\item{data}{A matching-prepared loanbook or ald file}
}
\value{
A matching-prepared file with simplified name.
}
\description{
This function is a wrapper for \code{replace_customer_name} to simplify the name
column of the matching-prepared loanbook or ald file.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  bridge_sector() \%>\%
  prepare_loanbook_for_matching() \%>\%
  simplify_name_column()

loanbook_demo \%>\%
  bridge_sector() \%>\%
  prepare_loanbook_for_matching() \%>\%
  overwrite_name_sector(overwrite_demo) \%>\%
  simplify_name_column()
}
\seealso{
\link{replace_customer_name} name simplification function.
}
