% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_name.R
\name{simplifyName}
\alias{simplifyName}
\title{Simplify customer name}
\usage{
simplifyName(name, cut.ownership = FALSE,
  reduction = pacta::data.name.reductions,
  ownership = pacta::data.ownership.types)
}
\arguments{
\item{name}{text string to simplify}

\item{cut.ownership}{flag that defines whether ownership type (like llc)
should be cut-off from name during simplification}

\item{reduction}{data frame with reduction rules to be applied, contains
columns From (for initial values) and To (for resulting values)}

\item{ownership}{vector of company ownership types to be disctinguished
for cut-off or separation}
}
\value{
Character string with simplified name
}
\description{
Takes customer name and simplifies it by
converting to lower case; converting to latin-ascii characters;
converting ownership types to standard abbreviations.
}
\examples{
simplifyName("Acuity Brands Inc")
simplifyName(c("3M Company", "Abbott Laboratories", "AbbVie Inc.", "Accenture plc"))
}
