% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquify_id_column.R
\name{uniquify_id_column}
\alias{uniquify_id_column}
\title{Overwrite \code{id_direct_loantaker} and \code{id_ultimate_parent} with unique values}
\usage{
uniquify_id_column(data, id_column, prefix)
}
\arguments{
\item{data}{A loanbook dataframe.}

\item{id_column}{A String giving the name of an \code{id_} column.}

\item{prefix}{A string giving a prefix for the values of the \code{id_column}.}
}
\value{
A loanbook dataframe with adjusted ids.
}
\description{
Given a loanbook dataframe, this function overwrites columns
\code{id_direct_loantaker} and \code{id_ultimate_parent} of a loanbook dataframe to
generate values that are unique by every combination of the columns
\code{name_ultimate_parent}, \code{name_direct_loantaker} and
\code{sector_classification_direct_loantaker}.
}
\examples{
library(dplyr)
library(r2dii.dataraw)

loanbook_demo \%>\%
  select(id_direct_loantaker, everything()) \%>\%
  # To more clearly show the effect of uniquify_id_column()
  mutate(id_ultimate_parent = "anything")

loanbook_demo \%>\%
  select(id_direct_loantaker, everything()) \%>\%
  uniquify_id_column(id_column = "id_direct_loantaker", prefix = "C")

# Same
loanbook_demo \%>\%
  select(id_ultimate_parent, everything()) \%>\%
  # To more clearly show the effect of uniquify_id_column()
  mutate(id_ultimate_parent = "anything")

loanbook_demo \%>\%
  select(id_ultimate_parent, everything()) \%>\%
  uniquify_id_column(id_column = "id_ultimate_parent", prefix = "C")
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description},
\link[r2dii.dataraw:loanbook_demo]{r2dii.dataraw::loanbook_demo}.

Other internal-ish: 
\code{\link{bridge_sector}()},
\code{\link{match_all_against_all}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_similarity}()},
\code{\link{to_alias}()}
}
\concept{internal-ish}
