% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_sector.R
\name{bridge_sector}
\alias{bridge_sector}
\title{Bridge classification between portfolio's system and 2dii's standard}
\usage{
bridge_sector(data)
}
\arguments{
\item{data}{A loanbook dataframe.}
}
\value{
A loanbook dataframe with additional \code{sector} and \code{borderline}
columns.
}
\description{
This function adds two columns, \code{sector} and \code{borderline} to the input
portfolio, corresponding to the bridged sector classification.
}
\details{
The r2dii.dataraw package must be attached (i.e. run \code{library(r2dii.dataraw)}
before you call \code{bridge_sector()}).
}
\examples{
# Must be attached
library(r2dii.dataraw)

bridge_sector(r2dii.dataraw::loanbook_demo)
}
\seealso{
Other functions to prepare data for matching.: 
\code{\link{overwrite_name_sector}()},
\code{\link{prepare_loanbook_for_matching}()},
\code{\link{pull_unique_name_sector_combinations}()}
}
\concept{functions to prepare data for matching.}
