% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_id.R
\name{gen_id}
\alias{gen_id}
\title{Populate \code{id_direct_loantaker}, and \code{id_ultimate_parent} in the loanbook}
\usage{
gen_id(data)
}
\arguments{
\item{data}{The input loanbook. See r2ii.dataraw for details.}
}
\value{
The input loanbook, with adjusted ids
}
\description{
This function generates identifiers for \code{id_direct_loantaker} and
\code{id_ultimate_parent} that are unique by every Name-Sector pair.
}
\examples{
lbk <- r2dii.dataraw::loanbook_demo
lbk_id_adjusted <- gen_id(lbk)
lbk_id_adjusted
}
