% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_by_loantaker_sector.R
\name{id_by_loantaker_sector}
\alias{id_by_loantaker_sector}
\title{Overwrite \code{id_direct_loantaker} and \code{id_ultimate_parent} with unique values}
\usage{
id_by_loantaker_sector(data)
}
\arguments{
\item{data}{A loanbook dataframe.}
}
\value{
A loanbook dataframe with adjusted ids.
}
\description{
Given a loanbook dataframe, this function overwrites columns
\code{id_direct_loantaker} and \code{id_ultimate_parent} of a loanbook dataframe to
generate values that are unique by every combination of the columns
\code{name_ultimate_parent}, \code{name_direct_loantaker} and
\code{sector_classification_direct_loantaker}.
}
\examples{
id_by_loantaker_sector(r2dii.dataraw::loanbook_demo)
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description},
\link[r2dii.dataraw:loanbook_demo]{r2dii.dataraw::loanbook_demo}.

Other internal-ish: 
\code{\link{bridge_sector}()},
\code{\link{replace_customer_name}()},
\code{\link{string_similarity}()}
}
\concept{internal-ish}
