% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_all_against_all.R
\name{match_all_against_all}
\alias{match_all_against_all}
\title{Score similarity between \code{simpler_name} values in two dataframes, by sector}
\usage{
match_all_against_all(x, y, group_by_sector = TRUE)
}
\arguments{
\item{x, y}{Dataframes}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Apply \code{string_similarity()} to all combinations of \code{simpler_name} values, by
sector, from two dataframes.
}
\examples{
library(dplyr)

x <- tibble(sector = c("A", "B", "B"), simpler_name = c("xa", "xb", "xc"))
y <- tibble(sector = c("A", "B", "C"), simpler_name = c("ya", "yb", "yc"))

scores <- match_all_against_all(x, y)
scores

scores \%>\%
  dplyr::filter(score > 0.5)
}
