% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_name.R
\name{match_name}
\alias{match_name}
\title{Match a loanbook (lbk) and asset-level datasets (ald) by the \verb{name_*} columns}
\usage{
match_name(
  loanbook,
  ald,
  by_sector = TRUE,
  min_score = 0.8,
  method = "jw",
  p = 0.1,
  overwrite = NULL
)
}
\arguments{
\item{loanbook}{Dataframes with \code{alias} and optionally \code{sector}
columns.}

\item{ald}{Dataframes with \code{alias} and optionally \code{sector}
columns.}

\item{by_sector}{Should the combinations be done by sector?}

\item{min_score}{A number (length-1) to set the minimum \code{score} values you
want to pick.}

\item{method}{Method for distance calculation. The default is \code{"osa"},
see \code{\link[stringdist]{stringdist-metrics}}.}

\item{p}{Penalty factor for Jaro-Winkler distance. The valid range for 
\code{p} is \code{0 <= p <= 0.25}. If \code{p=0} (default), the
Jaro-distance is returned. Applies only to \code{method='jw'}.}

\item{overwrite}{A dataframe used to overwrite the \code{sector} and/or \code{name}
columns of a particular direct loantaker or ultimate parent. To overwrite
only \code{sector}, the value in the \code{name} column should be \code{NA}.}
}
\value{
A dataframe with the same columns as the loanbook data with
additional columns: \code{id_lkb}, \code{sector_lbk}, \code{sector_ald}, \code{source_lbk},
\code{alias_lbk}, \code{alias_ald}, \code{score}, \code{name_ald}.
}
\description{
\code{match_name()} scores the match between names in a loanbook dataset (columns
\code{name_direct_loantaker} and \code{name_ultimate_parent}) with names in an
asset-level dataset (column \code{name_company}). The raw names are first
transformed and stored in a \code{alias} column, then the similarity between
the \code{alias} columns in each of the loanbook and ald datasets is scored
using \code{\link[stringdist:stringsim]{stringdist::stringsim()}}.
}
\section{Assigning aliases}{

The process to assign an alias for a \verb{name_*} column (i.e. the process to
create the \verb{alias_*} columns) applies best practices
commonly used in name matching algorithms:
\itemize{
\item Remove special characters.
\item Replace language specific characters.
\item Abbreviate certain names to reduce their importance in the matching.
\item Spell out numbers to increase their importance.
}
}

\examples{
library(dplyr)
library(r2dii.dataraw)

match_name(loanbook_demo, ald_demo)

match_name(
  loanbook_demo, ald_demo,
  min_score = 0.9,
  by_sector = FALSE
)
}
\seealso{
Other internal-ish: 
\code{\link{bridge_sector}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_similarity}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{user-oriented}
