% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ald_for_matching.R
\name{prepare_ald_for_matching}
\alias{prepare_ald_for_matching}
\title{A convenient shortcut to prepare ald data for matching}
\usage{
prepare_ald_for_matching(data)
}
\arguments{
\item{data}{A dataframe structured like \link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.}
}
\value{
A dataframe prepared for matching.
}
\description{
This function is a convenient shortcut for the most common steps for
preparing the ald dataset for matching.
}
\examples{
library(r2dii.dataraw)

prepare_ald_for_matching(ald_demo)
}
\seealso{
\link[r2dii.dataraw:ald_description]{r2dii.dataraw::ald_description}.

Other functions to prepare data for matching.: 
\code{\link{bridge_sector}()},
\code{\link{overwrite_name_sector}()},
\code{\link{prepare_loanbook_for_matching}()},
\code{\link{pull_unique_name_sector_combinations_ald}()},
\code{\link{pull_unique_name_sector_combinations}()}
}
\concept{functions to prepare data for matching.}
