% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{prepare_loanbook_for_matching}
\alias{prepare_loanbook_for_matching}
\title{Prepare a loanbook dataset for the fuzzy matching process}
\usage{
prepare_loanbook_for_matching(data, overwrite = NULL)
}
\arguments{
\item{data}{A dataframe. Should be a loanbook.}

\item{overwrite}{A dataframe used to overwrite the \code{sector} and/or \code{name}
columns of a particular direct loantaker or ultimate parent. To overwrite
only \code{sector}, the value in the \code{name} column should be \code{NA}.}
}
\value{
A dataframe with unique combinations of \code{name} + \code{sector}, including
all IDs, and with elements already manually overwritten.
}
\description{
This function prepares a loanbook dataset for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  prepare_loanbook_for_matching()

loanbook_demo \%>\%
  prepare_loanbook_for_matching(overwrite = overwrite_demo)
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description}

Other user-oriented: 
\code{\link{match_all_against_all}()},
\code{\link{prepare_ald_for_matching}()}
}
\concept{user-oriented}
