% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{prepare_loanbook_for_matching}
\alias{prepare_loanbook_for_matching}
\title{Prepare the loanbook elements to be matched.}
\usage{
prepare_loanbook_for_matching(data, overwrite = NULL)
}
\arguments{
\item{data}{A sector bridged loanbook dataframe.}

\item{overwrite}{A dataframe used to overwrite the sector and/or name of a
particular loan, direct loantaker or ultimate parent. Elements to be
overwritten are identified by any one of their IDs (loan, direct loantaker
or ultimate parent).}
}
\value{
A dataframe of all unique name + sector combinations, including all
IDs, and with elements already manually overwritten.
}
\description{
This function takes all unique name + sector combinations of a loanbook,
preparing it for the fuzzy matching process.
}
\examples{
library(r2dii.dataraw)

loanbook_demo \%>\%
  bridge_sector() \%>\%
  prepare_loanbook_for_matching()
}
