% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_loanbook_for_matching.R
\name{prepare_loanbook_for_matching}
\alias{prepare_loanbook_for_matching}
\title{A convenient shortcut to prepare loanbook data for matching}
\usage{
prepare_loanbook_for_matching(data, overwrite = NULL)
}
\arguments{
\item{data}{A dataframe structured like the output of \code{\link[=bridge_sector]{bridge_sector()}}.}

\item{overwrite}{A dataframe used to overwrite the sector and/or name of a
particular direct loantaker or ultimate parent. If only name (sector)
should be overwritten leave sector (name) as \code{NA}.}
}
\value{
A dataframe prepared for matching.
}
\description{
This function is a convenient shortcut for the most common steps for
preparing a loanbook dataset for matching.
}
\examples{
library(r2dii.dataraw)

prepare_loanbook_for_matching(loanbook_demo)

loanbook_demo \%>\%
  prepare_loanbook_for_matching(overwrite = overwrite_demo)
}
\seealso{
\link[r2dii.dataraw:loanbook_description]{r2dii.dataraw::loanbook_description}.

Other functions to prepare data for matching.: 
\code{\link{bridge_sector}()},
\code{\link{overwrite_name_sector}()},
\code{\link{pull_unique_name_sector_combinations}()}
}
\concept{functions to prepare data for matching.}
