% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize_at}
\alias{prioritize_at}
\title{Pick rows from a dataframe based on a priority set at some columns}
\usage{
prioritize_at(data, .at, priority = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{.at}{Most commonly, a character vector of one column name. For more
general usage see the \code{.vars} argument to \code{\link[dplyr:arrange_at]{dplyr::arrange_at()}}.}

\item{priority}{Most commonly, a character vector of the priority to
re-order the column(x) given by \code{.at}.}
}
\value{
A dataframe, commonly with less rows than the input.
}
\description{
Pick rows from a dataframe based on a priority set at some columns
}
\examples{
library(dplyr)

# styler: off
data <- tibble::tribble(
  ~x, ~y,
  1, "a",
  2, "a",
  2, "z",
)
# styler: on

data \%>\% prioritize_at("y")

data \%>\%
  group_by(x) \%>\%
  prioritize_at("y")

data \%>\%
  group_by(x) \%>\%
  prioritize_at(.at = "y", priority = c("z", "a")) \%>\%
  arrange(x) \%>\%
  ungroup()
}
\seealso{
Other internal-ish: 
\code{\link{add_sector_and_borderline}()},
\code{\link{prioritize_level}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_alias_similarity}()},
\code{\link{score_string_similarity}()},
\code{\link{select_chr}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
