% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize_level}
\alias{prioritize_level}
\title{Arrange unique \code{level} values in default order of \code{priority}}
\usage{
prioritize_level(data)
}
\arguments{
\item{data}{A dataframe, commonly the output of \code{\link[=match_name]{match_name()}}.}
}
\value{
A character vector of the default level priority per loan.
}
\description{
Arrange unique \code{level} values in default order of \code{priority}
}
\examples{
matched <- tibble::tibble(
  level = c(
    "intermediate_parent_1",
    "direct_loantaker",
    "direct_loantaker",
    "direct_loantaker",
    "ultimate_parent",
    "intermediate_parent_2"
  )
)
prioritize_level(matched)
}
\seealso{
Other internal-ish: 
\code{\link{add_sector_and_borderline}()},
\code{\link{prioritize_at}()},
\code{\link{restructure_ald_for_matching}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_alias_similarity}()},
\code{\link{score_string_similarity}()},
\code{\link{select_chr}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
