% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{prioritize_level}
\alias{prioritize_level}
\title{Arrange unique \code{level} values in default order of \code{priority}}
\usage{
prioritize_level(data)
}
\arguments{
\item{data}{A data frame, commonly the output of \code{\link[=match_name]{match_name()}}.}
}
\value{
A character vector of the default level priority per loan.
}
\description{
Arrange unique \code{level} values in default order of \code{priority}
}
\examples{
matched <- tibble::tibble(
  level = c(
    "intermediate_parent_1",
    "direct_loantaker",
    "direct_loantaker",
    "direct_loantaker",
    "ultimate_parent",
    "intermediate_parent_2"
  )
)
prioritize_level(matched)
}
