% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.match-package.R
\docType{package}
\name{r2dii.match-package}
\alias{r2dii.match}
\alias{r2dii.match-package}
\title{r2dii.match: Match Loanbook with Asset Level Data}
\description{
Match loanbook with asset level data. For more
    information visit <https://2degrees-investing.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.match}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.match/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (0000-0002-1986-7988) [contractor]

Authors:
\itemize{
  \item Klaus Hagedorn \email{klaus@2degrees-investing.org}
  \item Evgeny Petrovsky [contractor]
}

Other contributors:
\itemize{
  \item Jackson Hoffart \email{jackson@2degrees-investing.org} (0000-0002-8600-5042) [contributor]
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
