% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.match-package.R
\docType{package}
\name{r2dii.match-package}
\alias{r2dii.match}
\alias{r2dii.match-package}
\title{r2dii.match: Tools to Match Financial Portfolios with Climate Data}
\description{
These tools implement in R a fundamental part of
    the software 'PACTA' (Paris Agreement Capital Transition Assessment),
    which is a free tool that calculates the alignment between financial
    portfolios and climate scenarios (<https://2degrees-investing.org/>).
    Financial institutions use 'PACTA' to study how their capital
    allocation impacts the climate. This package matches data from
    financial portfolios to asset level data from market-intelligence
    databases (e.g. power plant capacities, emission factors, etc.). This
    is the first step to assess if a financial portfolio aligns with
    climate goals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/r2dii.match}
  \item \url{https://github.com/2DegreesInvesting/r2dii.match}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.match/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson@2degrees-investing.org} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID})
  \item Klaus Hagedorn \email{klaus@2degrees-investing.org}
  \item Florence Palandri \email{florence@2degrees-investing.org}
  \item Evgeny Petrovsky
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
