% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_loanbook_for_matching.R
\name{restructure_ald_for_matching}
\alias{restructure_ald_for_matching}
\title{Restructure an asset-level dataset (ald) in preparation for fuzzy matching}
\usage{
restructure_ald_for_matching(data)
}
\arguments{
\item{data}{A dataframe. Should be an asset-level dataset.}
}
\value{
A dataframe with unique combinations of \code{name} + \code{sector}, including
all IDs, and with elements already manually overwritten.
}
\description{
This function restructures an asset-level dataset (ald) in preparation for
the fuzzy matching process. Most notably, it outputs an the \code{alias} column
from values in the \code{name_company} column.
}
\examples{
library(r2dii.dataraw)

ald_demo \%>\%
  restructure_ald_for_matching()
}
\seealso{
\link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo} \code{\link[=to_alias]{to_alias()}}.

Other internal-ish: 
\code{\link{bridge_sector}()},
\code{\link{match_all_against_all}()},
\code{\link{restructure_loanbook_for_matching}()},
\code{\link{score_similarity}()},
\code{\link{to_alias}()},
\code{\link{uniquify_id_column}()}
}
\concept{internal-ish}
