% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_bridge.R
\name{sector_bridge}
\alias{sector_bridge}
\title{Bridge classification between portfolio's system and 2dii's standard}
\usage{
sector_bridge(data)
}
\arguments{
\item{data}{A loanbook dataframe.}
}
\value{
A loanbook dataframe with additional \code{sector} and \code{borderline}
columns.
}
\description{
This function adds two columns, \code{sector} and \code{borderline} to the input
portfolio, corresponding to the bridged sector classification.
}
\examples{
# Must be attached
library(r2dii.dataraw)

sector_bridge(r2dii.dataraw::loanbook_demo)
}
